openai_agent_sdk_realted/pyproject.toml
[project]
name = "openai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.1",
    "litellm==1.73.6",
    "openai-agents[litellm]>=0.1.0",
    "python-dotenv>=1.1.1",
]
